requirements.txt
aiofiles==23.2.1
aiosqlite==0.19.0
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
beautifulsoup4==4.12.2
billiard==4.2.4
cachetools==5.3.2
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
itsdangerous==2.1.2
kombu==5.6.1
libmagic==1.0
lxml==4.9.3
Mako==1.3.10
Markdown==3.5.1
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.24.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
